Mopho 2: Created rndm_0_comp_AllowChoiceToggles_Base.h and .cpp and declared and defined class AllowChoiceToggles_Base. In buttonClicked() added a condition to only respond to ALT-clicks if there are more than two columns. Renamed class AllowChoiceToggles to AllowChoiceToggles_Standard and made it inherit from AllowChoiceToggles_Base. Created rndm_1_comp_AllowChoiceToggles_PulseWidth.h and .cpp and declared and defined class AllowChoiceToggles_PulseWidth. Created rndm_1_comp_AllowChoiceToggles_LFO_Freq_Unsynced.h and .cpp and declared and defined class AllowChoiceToggles_LFO_Freq_Unsynced. Created rndm_1_comp_AllowChoiceToggles_LFO_Freq_Pitched.h and .cpp and declared and defined class AllowChoiceToggles_LFO_Freq_Pitched. Created rndm_1_comp_AllowChoiceToggles_LFO_Freq_Synced.h and .cpp and declared and defined class AllowChoiceToggles_LFO_Freq_Synced. Renamed rndm_1_gui_layer_AllowedChoices.h and .cpp to rndm_2_gui_layer_AllowedChoices_Standard.h and .cpp and renamed class GUI_Layer_AllowedChoices to GUI_Layer_AllowedChoices_Standard. Renamed rndm_1_gui_layer_AllowedChoices_OscShape.h and .cpp to rndm_2_gui_layer_AllowedChoices_OscShape.h and .cpp. In paint(), removed the code for outlining different control types. Declared Identifiers component_RedToggle_AllowLFO_Freq_Pitched, component_RedToggle_AllowLFO_Freq_Synced, component_RedToggle_AllowLFO_Freq_Unsynced, component_ToggleAllow_LFO_Freq_Pitched_, component_ToggleAllow_LFO_Freq_Synced_, and component_ToggleAllow_LFO_Freq_Unsynced_. Created rndm_2_gui_layer_AllowedChoices_LFO_Freq.h and .cpp and declared and defined class GUI_Layer_AllowedChoices_LFO_Freq.

+++ MophoGUI/MophoGUI 2.0/Source/unexposedParameters/up_tree_RandomizationOptions.cpp
@@ -473,11 +473,11 @@ void RandomizationOptions::allowAllChoicesForLFO_FreqParam(uint8 paramIndex) {
 	jassert(info.allowedChoicesTypeFor(paramIndex) == AllowedChoicesType::lfoFreq);
 	for (auto category = (int)Category::unsynced; category <= (int)Category::synced; ++category)
 		setCategoryIsAllowedForLFO_FreqParam(Category{ category }, true, paramIndex);
-	for (auto unsyncedFreq = 0; unsyncedFreq < EP::numberOfUnsyncedLFO_Frequencies; ++unsyncedFreq)
+	for (auto unsyncedFreq = (uint8)0; unsyncedFreq < EP::numberOfUnsyncedLFO_Frequencies; ++unsyncedFreq)
 		setUnsyncedFreqIsAllowedForLFO_FreqParam(unsyncedFreq, true, paramIndex);
-	for (auto pitchedFreq = 0; pitchedFreq < EP::numberOfPitchedLFO_Frequencies; ++pitchedFreq)
+	for (auto pitchedFreq = (uint8)0; pitchedFreq < EP::numberOfPitchedLFO_Frequencies; ++pitchedFreq)
 		setPitchedFreqIsAllowedForLFO_FreqParam(pitchedFreq, true, paramIndex);
-	for (auto syncedFreq = 0; syncedFreq < EP::numberOfSyncedLFO_Frequencies; ++syncedFreq)
+	for (auto syncedFreq = (uint8)0; syncedFreq < EP::numberOfSyncedLFO_Frequencies; ++syncedFreq)
 		setSyncedFreqIsAllowedForLFO_FreqParam(syncedFreq, true, paramIndex);
 }
